refactor: extract scenario wizard scene flow helpers
